Fix utils::is_public
